Fix Kolors, Cascade, HiDream, Z-Image backend bugs
- Kolors: permute ChatGLM hidden states from [seq, batch, hidden]
  to [batch, seq, hidden]. Without this the UNet receives transposed
  text conditioning, producing garbage training for batch_size > 1.

- Cascade: override training_step with correct StableCascadeUNet
  forward signature. The base training_step passed wrong positional
  args (integer timesteps instead of ratios, encoder_hidden as
  clip_text_pooled, added_cond_kwargs not accepted). Training was
  completely broken.

- HiDream: pass T5 and Llama hidden states as separate kwargs
  (encoder_hidden_states_t5, encoder_hidden_states_llama3) instead
  of concatenating all 4 encoders into one tensor. Fix pooled kwarg
  name from pooled_projections to pooled_embeds. Previously the
  transformer received corrupted text conditioning and silently
  dropped all pooled CLIP embeddings.

- Z-Image: cast t to float32 before computing debiased estimation
  weight, matching the base class behavior. Prevents catastrophic
  cancellation near t=1.0 in bf16.

+++ dataset_sorter/train_backend_zimage.py
@@ -851,7 +851,7 @@ class ZImageBackend(TrainBackendBase):
 
         if config.debiased_estimation:
             # Clamp denominator to prevent extreme weights when t → 1.0
-            weight = 1.0 / torch.clamp(1.0 - t + 1e-6, min=0.01)
+            weight = 1.0 / torch.clamp(1.0 - t.float() + 1e-6, min=0.01)
             loss = loss * weight
 
         # min_snr_gamma: not applicable to flow matching (requires alphas_cumprod)
